added props to exam schedule
