+++ packages/alea-frontend/pages/course-home/[courseId].tsx
@@ -47,6 +47,7 @@ import { RecordedSyllabus } from '../../components/RecordedSyllabus';
 import { useStudentCount } from '../../hooks/useStudentCount';
 import { getLocaleObject } from '../../lang/utils';
 import MainLayout from '../../layouts/MainLayout';
+import { ExamSchedule } from '../../components/ExamSchedule';
 
 export function getCourseEnrollmentAcl(courseId: string, instanceId: string) {
   return `${courseId}-${instanceId}-enrollments`;
@@ -73,67 +74,6 @@ export async function handleEnrollment(userId: string, courseId: string, current
   }
 }
 
-export function ExamSchedule({ examDates }) {
-  if (!examDates?.length) return null;
-
-  const timeZone = 'Europe/Berlin';
-
-  return (
-    <Box
-      sx={{
-        px: { xs: 1, sm: 2 },
-        py: { xs: 1, sm: 1.5 },
-        mt: 2,
-        borderRadius: '8px',
-        background: 'linear-gradient(135deg, #fff3e0, #ffe0b2)',
-        border: '1px solid #ffcc80',
-      }}
-    >
-      <Box sx={{ display: 'flex', alignItems: 'center', gap: 1.5, mb: 1.5 }}>
-        <QuizIcon sx={{ color: '#bf360c', fontSize: '20px' }} />
-        <Typography variant="h6" sx={{ fontWeight: 600, color: '#bf360c', fontSize: '1rem' }}>
-          Exam Schedule
-        </Typography>
-      </Box>
-
-      {examDates.map((exam, idx) => {
-        const hasTime = exam.examStartTime && exam.examEndTime;
-        const start = new Date(`${exam.examDate}T${exam.examStartTime || '00:00'}:00+02:00`);  //+02:00 is used because we assume Europe/Berlin time.
-        const end = new Date(`${exam.examDate}T${exam.examEndTime || '00:00'}:00+02:00`);
-
-        const formattedDate = start.toLocaleDateString('en-GB', {
-          day: '2-digit',
-          month: 'short',
-          year: 'numeric',
-          timeZone,
-        });
-
-        const formattedStart = hasTime
-          ? start.toLocaleTimeString('en-GB', {
-              hour: '2-digit',
-              minute: '2-digit',
-              timeZone,
-            })
-          : '';
-
-        const formattedEnd = hasTime
-          ? end.toLocaleTimeString('en-GB', {
-              hour: '2-digit',
-              minute: '2-digit',
-              timeZone,
-            })
-          : '';
-
-        return (
-          <Typography key={idx} variant="body2" sx={{ color: '#e65100', mb: 1, fontWeight: 500 }}>
-            🗓 {formattedDate}
-            {hasTime && ` ⏰ ${formattedStart} – ${formattedEnd} (${timeZone})`}
-          </Typography>
-        );
-      })}
-    </Box>
-  );
-}
 function CourseComponentLink({ href, children, sx }: { href: string; children: any; sx?: any }) {
   return (
     <Link href={href}>
@@ -241,7 +181,8 @@ function CourseScheduleSection({
     [courseId]
   );
 
-  const examDates = semesterPeriods[CURRENT_TERM]?.examDates.filter((e) => e.courseId === courseId) || [];
+  const examDates =
+    semesterPeriods[CURRENT_TERM]?.examDates.filter((e) => e.courseId === courseId) || [];
 
   useEffect(() => {
     async function fetchNextLectureDates() {
@@ -375,8 +316,7 @@ function CourseScheduleSection({
           </Box>
         )}
 
-       <ExamSchedule examDates={examDates} />
-
+        <ExamSchedule examDates={examDates} />
 
         {userId && (
           <PersonalCalendarSection
